+++ tybotRoute/CHANGELOG.md
@@ -5,8 +5,18 @@
 available on:
  ▶️ https://www.npmjs.com/package/@tiledesk/tiledesk-tybot-connector
 
+# v0.2.37
+- improved GptTask action
+- improved AskGPT action 
+
+# v0.2.36
+- Fixes json conversion in GptTask issue
+
 # v0.2.35
 - added validation check for requests ids (support-group- && automation-request-)
+- refactorized GptTask, AskGPT and Qapla actions in order to support integrations
+- added integrated condition in GptTask 
+- added automatic JSON conversion in GptTask
 
 # v0.2.34
 - stopOnConditionMet is always true in DirJSONCondition